requirements/prod.txt
Django==4.2.5
psycopg2-binary==2.9.7
djangorestframework==3.14.0
PyJWT==2.8.0
django-cors-headers==4.2.0
django-import-export==3.3.1
pandas
requests
celery
redis
django-celery-results
django-celery-beat
django-storages